Use GMT times for ICS values
Offset start and end date by the time zone offset so that the hours, etc. will be in UTC instead of local time. Note that the timezone of the dates is unaffected - we are simply offsetting by the time zone offset so that we can get the year, month, days, hours, and minutes from the equivalent UTC time.